+++ assets/l10n/intl_en.arb
@@ -234,10 +234,6 @@
   "@goToReceipts": {},
   "goToListings": "go to listings",
   "@goToListings": {},
-  "changePassword": "change password",
-  "@changePassword": {},
-  "deleteUser": "delete user",
-  "@deleteUser": {},
   "clickLocation": "tap here to select location",
   "@clickLocation": {},
   "listReceipt": "list of receipts",
